Add files via upload
worked on the user interface 

design of the project. .she focuses on what users may have to be compelled to do and ensuring 

that the interface has components that are straightforward to access, understand, and use to 

facilitate those actions.that includes Home page, food menu categories,food ordering 

process and customers account details.